+++ lib/minitest-activemodel/matchers/validate_uniqueness_matcher.rb
@@ -5,10 +5,13 @@ module MiniTest
       #
       # Options:
       # * <tt>scoped_to</tt> - field(s) to scope the uniqueness to.
+      # * <tt>case_insensitive</tt> - ensures that the validation doesn't
+      #   check case. Off by default. Ignored by non-text attributes.
       #
       #   it { must validate_uniqueness_of(:email) }
       #   it { must validate_uniqueness_of(:email).scoped_to(:name) }
       #   it { must validate_uniqueness_of(:email).scoped_to(:first_name, :last_name) }
+      #   it { must validate_uniqueness_of(:keyword).case_insensitive }
       def validate_uniqueness_of attr
         ValidateUniquenessMatcher.new attr
       end
@@ -18,6 +21,7 @@ module MiniTest
       class ValidateUniquenessMatcher < ValidationMatcher
         def initialize attr
           @scope = nil
+          @case_insensitive = false
 
           super attr, :uniqueness
         end
@@ -27,10 +31,16 @@ module MiniTest
           self
         end
 
+        def case_insensitive
+          @case_insensitive = true
+          self
+        end
+
         def matches? subject
           return false unless @result = super(subject)
 
           check_scope if @scope
+          check_case_sensivity if @case_insensitive
 
           @result
         end
@@ -38,6 +48,7 @@ module MiniTest
         def description
           desc = []
           desc  << " scoped to #{to_sentence(@scope)}" if @scope
+          desc << ' allowing case insensitive values'  if @case_insensitive
           super << desc.to_sentence
         end
 
@@ -51,6 +62,15 @@ module MiniTest
             @positive_message << message
           else
             @negative_message << message
+            @result = false
+          end
+        end
+
+        def check_case_sensivity
+          if @validator.options[:case_sensitive] == false
+            @positive_message << ' with case insensitive values'
+          else
+            @negative_message << ' without case insensitive values'
             @result = false
           end
         end