+++ test/cases/validate_uniqueness_matcher_test.rb
@@ -18,4 +18,8 @@ class ValidateUniquenessMatcherTest < MiniTest::Unit::TestCase
   test 'must not validate scope of an unique attribute' do
     assert_wont validate_uniqueness_of(:email).scoped_to(:none), Person
   end
+
+  test 'must validate uniqueness validation does not check case' do
+    assert_must validate_uniqueness_of(:email).case_insensitive, Person
+  end
 end