+++ test/models/person.rb
@@ -1,3 +1,3 @@
 class Person < ActiveRecord::Base
-  validates_uniqueness_of :email, scope: :site
+  validates_uniqueness_of :email, scope: :site, case_sensitive: false
 end