+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 + Add `validate_uniqueness_of` matcher.
 + Add `scoped_to` option to `validate_uniqueness_of` matcher.
++ Add `case_insensitive` option to `validate_uniqueness_of` matcher.
 
 # 0.0.1 - 10/13/12